rendering refactor
