--- engine/src/rendering/directx/dm_directx_device.h
@@ -1,18 +1,0 @@
-#ifndef __DM_DIRECTX_DEVICE_H__
-#define __DM_DIRECTX_DEVICE_H__
-
-#include "core/dm_defines.h"
-
-#ifdef DM_DIRECTX
-
-#include "dm_directx_renderer.h"
-
-bool dm_directx_create_device(dm_directx_renderer* renderer);
-void dm_directx_destroy_device(dm_directx_renderer* renderer);
-#if DM_DEBUG
-void dm_directx_device_report_live_objects(dm_directx_renderer* renderer);
-#endif
-
-#endif
-
-#endif
