--- engine/src/rendering/directx/dm_directx_rendertarget.c
@@ -1,37 +1,0 @@
-#include "dm_directx_rendertarget.h"
-
-#ifdef DM_DIRECTX
-
-#include "core/dm_assert.h"
-#include "core/dm_mem.h"
-
-bool dm_directx_create_rendertarget(dm_directx_renderer* renderer, dm_directx_pipeline* pipeline)
-{
-	DM_ASSERT_MSG(renderer->device, "DirectX device is NULL!");
-	DM_ASSERT_MSG(renderer->swap_chain, "DirectX swap chain is NULL!");
-
-	HRESULT hr;
-	ID3D11Device* device = renderer->device;
-	IDXGISwapChain* swap_chain = renderer->swap_chain;
-	
-	DX_ERROR_CHECK(swap_chain->lpVtbl->GetBuffer(swap_chain, 0, &IID_ID3D11Texture2D, (void**)&(ID3D11Resource*)pipeline->render_back_buffer), "IDXGISwapChain::GetBuffer failed!");
-	device->lpVtbl->CreateRenderTargetView(device, (ID3D11Resource*)pipeline->render_back_buffer, NULL, &pipeline->render_view);
-	dm_mem_db_adjust(sizeof(ID3D11Texture2D), DM_MEM_RENDER_PIPELINE, DM_MEM_ADJUST_ADD);
-	dm_mem_db_adjust(sizeof(ID3D11RenderTargetView), DM_MEM_RENDER_PIPELINE, DM_MEM_ADJUST_ADD);
-
-	return true;
-}
-
-void dm_directx_destroy_rendertarget(dm_directx_pipeline* pipeline)
-{
-	ID3D11RenderTargetView* view = pipeline->render_view;
-	ID3D11Texture2D* back_buffer = pipeline->render_back_buffer;
-
-	DX_RELEASE(view);
-	DX_RELEASE(back_buffer);
-
-	dm_mem_db_adjust(sizeof(ID3D11Texture2D), DM_MEM_RENDER_PIPELINE, DM_MEM_ADJUST_SUBTRACT);
-	dm_mem_db_adjust(sizeof(ID3D11RenderTargetView), DM_MEM_RENDER_PIPELINE, DM_MEM_ADJUST_SUBTRACT);
-}
-
-#endif
