+++ engine/src/rendering/dm_geometry.c
@@ -1,6 +1,7 @@
 #include "dm_geometry.h"
 
 #include "dm_renderer.h"
+#include "dm_renderer_api.h"
 
 bool dm_geometry_load_primitives()
 {
@@ -36,20 +37,20 @@ bool dm_geometry_load_primitives()
 		{ { 0.5f,  0.5f, -0.5f}, {1.0f, 0.0f, 0.0f}, {1.0f, 1.0f} },
 		{ { 0.5f,  0.5f,  0.5f}, {1.0f, 0.0f, 0.0f}, {0.0f, 1.0f} }
 	};
-
+    
 	dm_index_t cube_indices[] = {
-		 0,  1,  2,    2,  3,  0,
-		 4,  5,  6,    6,  7,  4,
-		 8,  9, 10,   10, 11,  8,
+        0,  1,  2,    2,  3,  0,
+        4,  5,  6,    6,  7,  4,
+        8,  9, 10,   10, 11,  8,
 		12, 13, 14,   14, 15, 12,
 		16, 17, 18,   18, 19, 16,
 		20, 21, 22,   22, 23, 20
 	};
-
+    
 	uint32_t num_vertices = sizeof(cube_vertices) / sizeof(dm_vertex_t);
 	uint32_t num_indices = sizeof(cube_indices) / sizeof(dm_index_t);
-
-	dm_renderer_submit_vertex_data(cube_vertices, cube_indices, num_vertices, num_indices, "cube");
-
+    
+	dm_renderer_api_submit_vertex_data("cube", cube_vertices, cube_indices, num_vertices, num_indices);
+    
 	return true;
 }
