--- engine/src/rendering/opengl/dm_opengl_shader.c
@@ -1,333 +1,0 @@
-#include "dm_opengl_shader.h"
-
-#ifdef DM_OPENGL
-
-#include "dm_opengl_enum_conversion.h"
-#include "core/dm_logger.h"
-#include "core/dm_assert.h"
-#include "core/dm_mem.h"
-
-GLuint dm_opengl_compile_shader(dm_shader_desc desc);
-bool dm_opengl_validate_shader(GLuint shader);
-bool dm_opengl_validate_program(GLuint program);
-
-bool dm_opengl_create_shader(dm_shader* shader)
-{
-    shader->internal_shader = dm_alloc(sizeof(dm_internal_shader), DM_MEM_RENDERER_SHADER);
-    dm_internal_shader* internal_shader = shader->internal_shader;
-
-    GLuint vertex_shader = dm_opengl_compile_shader(shader->vertex_desc);
-    GLuint frag_shader = dm_opengl_compile_shader(shader->pixel_desc);
-
-    DM_LOG_DEBUG("Linking shader...");
-    internal_shader->id = glCreateProgram();
-    glAttachShader(internal_shader->id, vertex_shader);
-    glCheckErrorReturn();
-    glAttachShader(internal_shader->id, frag_shader);
-    glCheckErrorReturn();
-    glLinkProgram(internal_shader->id);
-    glCheckErrorReturn();
-
-    if (!dm_opengl_validate_program(internal_shader->id))
-    {
-        DM_LOG_FATAL("Failed to validate OpenGL shader!");
-        return false;
-    }
-
-    glDetachShader(internal_shader->id, vertex_shader);
-    glCheckErrorReturn();
-    glDetachShader(internal_shader->id, frag_shader);
-    glCheckErrorReturn();
-
-    glDeleteShader(vertex_shader);
-    glCheckErrorReturn();
-    glDeleteShader(frag_shader);
-    glCheckErrorReturn();
-
-    return true;
-}
-
-void dm_opengl_delete_shader(dm_shader* shader)
-{
-    dm_internal_shader* internal_shader = shader->internal_shader;
-
-    glDeleteProgram(internal_shader->id);
-    glCheckError();
-    dm_free(shader->internal_shader, sizeof(dm_internal_shader), DM_MEM_RENDERER_SHADER);
-}
-
-void dm_opengl_bind_shader(dm_shader* shader)
-{
-    dm_internal_shader* internal_shader = shader->internal_shader;
-
-    glUseProgram(internal_shader->id);
-    glCheckError();
-}
-
-bool dm_opengl_create_uniform(dm_uniform* uniform, dm_shader* shader)
-{
-    uniform->internal_uniform = dm_alloc(sizeof(dm_internal_uniform), DM_MEM_RENDERER_SHADER);
-
-    dm_internal_uniform* internal_uni = uniform->internal_uniform;
-    dm_internal_shader* internal_shader = shader->internal_shader;
-
-    internal_uni->location = glGetUniformLocation(internal_shader->id, uniform->name);
-    if (internal_uni->location == -1)
-    {
-        DM_LOG_FATAL("Could not find uniform: '%s'", uniform->name);
-        return false;
-    }
-
-    return true;
-}
-
-void dm_opengl_destroy_uniform(dm_uniform* uniform)
-{
-    dm_free(uniform->internal_uniform, sizeof(dm_internal_uniform), DM_MEM_RENDERER_SHADER);
-}
-
-bool dm_opengl_bind_uniform(dm_uniform* uniform)
-{
-    dm_internal_uniform* internal_uniform = uniform->internal_uniform;
-
-    switch (uniform->desc.data_t)
-    {
-    case DM_UNIFORM_DATA_T_INT:
-    {
-        switch (uniform->desc.count)
-        {
-        case 1:
-        {
-            int data = *(int*)uniform->data;
-            glUniform1i(internal_uniform->location, data);
-            glCheckErrorReturn();
-        } break;
-        case 2:
-        {
-            dm_vec2 data = *(dm_vec2*)uniform->data;
-            glUniform2i(internal_uniform->location, data.x, data.y);
-            glCheckErrorReturn();
-        } break;
-        case 3:
-        {
-            dm_vec3 data = *(dm_vec3*)uniform->data;
-            glUniform3i(internal_uniform->location, data.x, data.y, data.z);
-            glCheckErrorReturn();
-        } break;
-        case 4:
-        {
-            dm_vec4 data = *(dm_vec4*)uniform->data;
-            glUniform4i(internal_uniform->location, data.x, data.y, data.z, data.w);
-            glCheckErrorReturn();
-        } break;
-        default:
-            DM_LOG_FATAL("Trying to upload to uniform with wrong size of ints");
-            return false;
-        }
-    } break;
-
-    case DM_UNIFORM_DATA_T_BOOL:
-    case DM_UNIFORM_DATA_T_UINT:
-    {
-        switch (uniform->desc.count)
-        {
-        case 1:
-        {
-            uint32_t data = *(uint32_t*)uniform->data;
-            glUniform1ui(internal_uniform->location, data);
-            glCheckErrorReturn();
-        } break;
-        case 2:
-        {
-            dm_vec2 data = *(dm_vec2*)uniform->data;
-            glUniform2ui(internal_uniform->location, data.x, data.y);
-            glCheckErrorReturn();
-        } break;
-        case 3:
-        {
-            dm_vec3 data = *(dm_vec3*)uniform->data;
-            glUniform3ui(internal_uniform->location, data.x, data.y, data.z);
-            glCheckErrorReturn();
-        } break;
-        case 4:
-        {
-            dm_vec4 data = *(dm_vec4*)uniform->data;
-            glUniform4ui(internal_uniform->location, data.x, data.y, data.z, data.w);
-            glCheckErrorReturn();
-        } break;
-        default:
-            DM_LOG_FATAL("Trying to upload to uniform with wrong size of uints");
-            return false;
-        }
-    } break;
-
-    case DM_UNIFORM_DATA_T_FLOAT:
-    {
-        switch (uniform->desc.count)
-        {
-        case 1:
-        {
-            float data = *(float*)uniform->data;
-            glUniform1f(internal_uniform->location, data);
-            glCheckErrorReturn();
-        } break;
-        case 2:
-        {
-            dm_vec2 data = *(dm_vec2*)uniform->data;
-            glUniform2f(internal_uniform->location, data.x, data.y);
-            glCheckErrorReturn();
-        } break;
-        case 3:
-        {
-            dm_vec3* data = (dm_vec3*)uniform->data;
-            glUniform3f(internal_uniform->location, data->x, data->y, data->z);
-            glCheckErrorReturn();
-        } break;
-        case 4:
-        {
-            dm_vec4 data = *(dm_vec4*)uniform->data;
-            glUniform4f(internal_uniform->location, data.x, data.y, data.z, data.w);
-            glCheckErrorReturn();
-        } break;
-        default:
-            DM_LOG_FATAL("Trying to upload to uniform with wrong size of floats");
-            return false;
-        }
-    } break;
-
-    case DM_UNIFORM_DATA_T_MATRIX_INT:
-    case DM_UNIFORM_DATA_T_MATRIX_FLOAT:
-    {
-        switch (uniform->desc.count)
-        {
-        case 2:
-        {
-            dm_mat2 data = *(dm_mat2*)uniform->data;
-            glUniformMatrix2fv(internal_uniform->location, 1, GL_FALSE, &data.m[0]);
-            glCheckErrorReturn();
-        } break;
-        case 3:
-        {
-            dm_mat3 data = *(dm_mat3*)uniform->data;
-            glUniformMatrix3fv(internal_uniform->location, 1, GL_FALSE, &data.m[0]);
-            glCheckErrorReturn();
-        } break;
-        case 4:
-        {
-            dm_mat4 data = *(dm_mat4*)uniform->data;
-            glUniformMatrix4fv(internal_uniform->location, 1, GL_FALSE, &data.m[0]);
-            glCheckErrorReturn();
-        } break;
-        default:
-            DM_LOG_FATAL("Trying to upload to uniform with wrong size of matrices");
-            return false;
-        }
-    } break;
-
-    default:
-        DM_LOG_FATAL("Unknown constant buffer data type!");
-        return false;
-    } 
-
-    return true;
-}
-
-/*
-uses dm_shader_desc to compile a glsl shader from file
-*/
-GLuint dm_opengl_compile_shader(dm_shader_desc desc)
-{
-    GLenum shader_type = dm_shader_to_opengl_shader(desc.type);
-    GLuint shader = glCreateShader(shader_type);
-    glCheckError();
-
-    DM_LOG_DEBUG("Compiling shader: %s", desc.path);
-
-    FILE* file = fopen(desc.path, "r");
-    DM_ASSERT_MSG(file, "Could not fopen file: %s", desc.path);
-
-    // determine size of memory to allocate to the buffer
-    fseek(file, 0, SEEK_END);
-    size_t length = ftell(file);
-    fseek(file, 0, SEEK_SET);
-    char* string = dm_alloc(length + 1, DM_MEM_STRING);
-    DM_ASSERT(string);
-
-    // instead of setting string[length] to nul, we instead
-    // find the last character that isn't garbage and set that
-    // to nul. otherwise there could be garbage in between and
-    // cause a seg fault
-    //fread(string, length, 1, file);
-    size_t read_count = fread(string, 1, length, file);
-    string[read_count] = '\0';
-    fclose(file);
-    DM_ASSERT(string);
-
-    const char* source = string;
-    GLint l = (GLint)length;
-    glShaderSource(shader, 1, &source, &l);
-    glCheckError();
-    glCompileShader(shader);
-    glCheckError();
-
-    DM_ASSERT(dm_opengl_validate_shader(shader));
-
-    dm_free(string, length+1, DM_MEM_STRING);
-
-    return shader;
-}
-
-bool dm_opengl_validate_shader(GLuint shader)
-{
-    GLint result = -1;
-    int length = -1;
-
-    glGetShaderiv(shader, GL_COMPILE_STATUS, &result);
-    glCheckErrorReturn();
-
-    if (result != GL_TRUE)
-    {
-        GLchar message[512];
-        glGetShaderiv(shader, GL_INFO_LOG_LENGTH, &length);
-        glCheckErrorReturn();
-        glGetShaderInfoLog(shader, sizeof(message), &length, message);
-        glCheckErrorReturn();
-        DM_LOG_FATAL("%s", message);
-
-        glDeleteShader(shader);
-        glCheckErrorReturn();
-
-        return false;
-    }
-
-    return true;
-}
-
-bool dm_opengl_validate_program(GLuint program)
-{
-    GLint result = -1;
-    int length = -1;
-
-    glGetProgramiv(program, GL_LINK_STATUS, &result);
-    glCheckErrorReturn();
-
-    if (result == GL_FALSE)
-    {
-        DM_LOG_ERROR("OpenGL Error: %d", glGetError());
-        char message[512];
-        glGetProgramiv(program, GL_INFO_LOG_LENGTH, &length);
-        glCheckErrorReturn();
-        glGetProgramInfoLog(program, length, NULL, message);
-        glCheckErrorReturn();
-        DM_LOG_FATAL("%s", message);
-
-        glDeleteProgram(program);
-        glCheckErrorReturn();
-
-        return false;
-    }
-
-    return true;
-}
-
-#endif
